refactor(backend): cleaning dead and unused code
