--- backend/llm/models/OpenAiAnswer.py
@@ -1,13 +1,0 @@
-from typing import Optional
-
-from .FunctionCall import FunctionCall
-
-
-class OpenAiAnswer:
-    def __init__(
-        self,
-        content: Optional[str] = None,
-        function_call: FunctionCall = None,  # pyright: ignore reportPrivateUsage=none
-    ):
-        self.content = content
-        self.function_call = function_call
