--- backend/llm/prompts/LANGUAGE_PROMPT.py
@@ -1,12 +1,0 @@
-from langchain.prompts.prompt import PromptTemplate
-
-prompt_template = """Your name is Quivr. You are a second brain. A person will ask you a question and you will provide a helpful answer. Write the answer in the same language as the question. If you don't know the answer, just say that you don't know. Don't try to make up an answer. Use the following context to answer the question:
-
-
-{context}
-
-Question: {question}
-Helpful Answer:"""
-QA_PROMPT = PromptTemplate(
-    template=prompt_template, input_variables=["context", "question"]
-)
